django/requirements.txt
Django==4.0.4; python_version >= '3.8'
Django==3.2.18; python_version < '3.8'
Pillow==9.2.0
django-tables2==2.4.1
django-crispy-forms==1.14.0
django-environ==0.10.0
psycopg2-binary==2.9.6
google-cloud-secret-manager==2.16.1
django-storages[google]==1.13.2
google-cloud-storage==2.7.0
